pyproject.toml
[project]
name = "really-aio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "argon2>=0.1.10",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.123.0",
    "jinja2>=3.1.6",
    "jinja2-fragments>=1.11.0",
    "ldap3>=2.9.1",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn>=0.38.0",
]
